+++ package.json
@@ -117,6 +117,7 @@
     "react-native-progress": "^3.6.0",
     "react-native-qrcode-svg": "5.1.2",
     "react-native-randombytes": "^3.5.3",
+    "react-native-reanimated": "^1.4.0",
     "react-native-screens": "^1.0.0-alpha.22",
     "react-native-sentry": "0.42.0",
     "react-native-simple-store": "2.0.2",
@@ -127,8 +128,9 @@
     "react-native-udp": "^2.6.0",
     "react-native-vector-icons": "^6.4.2",
     "react-native-version-number": "^0.3.6",
-    "react-navigation": "^3.11.1",
-    "react-navigation-tabs": "0.8.4",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3",
+    "react-navigation-tabs": "2.6.0",
     "react-redux": "5.1.1",
     "readable-stream": "1.0.33",
     "redux": "4.0.4",
