+++ src/screens/AddAccounts/03-Accounts.js
@@ -14,7 +14,7 @@ import { translate, Trans } from "react-i18next";
 import { StyleSheet, View } from "react-native";
 // $FlowFixMe
 import { SafeAreaView, ScrollView } from "react-navigation";
-import type { NavigationScreenProp } from "react-navigation";
+import type { NavigationStackProp } from "react-navigation-stack";
 import type { CryptoCurrency, Account } from "@ledgerhq/live-common/lib/types";
 import { getCurrencyBridge } from "@ledgerhq/live-common/lib/bridge";
 import { replaceAccounts } from "../../actions/accounts";
@@ -47,7 +47,7 @@ const SectionAccounts = ({ defaultSelected, ...rest }: *) => {
 };
 
 type Props = {
-  navigation: NavigationScreenProp<{
+  navigation: NavigationStackProp<{
     params: {
       currency: CryptoCurrency,
       deviceId: string,
