+++ src/screens/FirmwareUpdate/03-MCU.js
@@ -3,7 +3,7 @@
 import React, { Component } from "react";
 import { View, StyleSheet } from "react-native";
 import { SafeAreaView } from "react-navigation";
-import type { NavigationScreenProp } from "react-navigation";
+import type { NavigationStackProp } from "react-navigation-stack";
 import { translate, Trans } from "react-i18next";
 import firmwareUpdateMain from "@ledgerhq/live-common/lib/hw/firmwareUpdate-main";
 import type { FirmwareUpdateContext } from "@ledgerhq/live-common/lib/types/manager";
@@ -18,7 +18,7 @@ import Installing from "../../components/Installing";
 
 const forceInset = { bottom: "always" };
 
-type Navigation = NavigationScreenProp<{
+type Navigation = NavigationStackProp<{
   params: {
     deviceId: string,
     firmware: FirmwareUpdateContext,
