+++ src/screens/FirmwareUpdate/04-Failure.js
@@ -2,7 +2,7 @@
 import React, { Component } from "react";
 import { View, StyleSheet } from "react-native";
 import { SafeAreaView } from "react-navigation";
-import type { NavigationScreenProp } from "react-navigation";
+import type { NavigationStackProp } from "react-navigation-stack";
 import { translate, Trans } from "react-i18next";
 
 import { TrackScreen } from "../../analytics";
@@ -13,7 +13,7 @@ import colors from "../../colors";
 
 const forceInset = { bottom: "always" };
 
-type Navigation = NavigationScreenProp<{
+type Navigation = NavigationStackProp<{
   params: {
     deviceId: string,
     error: Error,
